add more examples
